pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bytecli"
version = "1.0.0"
description = "Local voice-to-text dictation tool for Ubuntu/Linux"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "python-xlib",
    "sounddevice",
    "pulsectl",
    "numpy",
    "dbus-python",
    "PyGObject",
]

[project.scripts]
bytecli-service = "bytecli.service.main:main"
bytecli-indicator = "bytecli.indicator.main:main"
bytecli-settings = "bytecli.settings.main:main"

[tool.setuptools.packages.find]
include = ["bytecli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
